2022/11/n8n-playground/pyproject.toml
[tool.poetry]
name = "random_backend"
version = "0.1.0"
description = ""
authors = ["Willian Antunes <willian.lima.antunes@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
Faker = "^15.3.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target_version = ["py310"]
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
    | settings.py
  )/
)
'''
